rename eject -> swizzle
