+++ packages/docusaurus/lib/index.js
@@ -8,12 +8,12 @@
 const build = require('./commands/build');
 const init = require('./commands/init');
 const start = require('./commands/start');
-const eject = require('./commands/eject');
+const swizzle = require('./commands/swizzle');
 const deploy = require('./commands/deploy');
 
 module.exports = {
   build,
-  eject,
+  swizzle,
   init,
   start,
   deploy,
